fix(atlas): marca-texto volta a ser translúcido, e o plano para de piscar ao mover
A opacidade nunca chegava ao banco. No INSERT da anotação, `NULLIF($7,0)`
comparava o parâmetro com o literal 0, que é inteiro: o Postgres inferia o
parâmetro como inteiro, 0.35 virava 0, caía no COALESCE e voltava 1. Todo
marca-texto era gravado opaco, tapando justamente a cota que ele deveria
realçar. A largura escapou por acaso, porque 12 é inteiro.

No leitor, a camada nítida cobre só o que está na tela, e ao arrastar a faixa
que entrava aparecia vazia até o redesenho chegar. Agora a folha inteira é
desenhada uma vez em resolução modesta e fica por baixo, acompanhando o gesto
só por CSS: o que chega já vem preenchido.

A camada nítida também passa a alternar entre dois canvas. O quadro novo é
desenhado escondido e só então vem para a frente, com um suspiro de transição
por cima do anterior. Não se vê mais o canvas sendo pintado, e a troca de
nitidez deixou de ser um salto.

+++ BOR2/apps/api/internal/handler/atlas.go
@@ -1428,10 +1428,13 @@ func (h *AtlasHandler) CreateAnnotation(c *fiber.Ctx) error {
 		in.ID = uuid.NewString()
 	}
 	userID, _ := actor(c)
+	// O cast para numeric não é enfeite: comparado com o literal 0, que é inteiro,
+	// o parâmetro era inferido como inteiro e toda opacidade fracionária virava
+	// zero, caía no COALESCE e voltava 1. O marca-texto saía opaco sempre.
 	_, err = h.db.Exec(c.Context(), `
 		INSERT INTO atlas_annotation (id, sheet_id, author_id, tool, color, width, opacity, geometry)
 		VALUES ($1,$2,$3,COALESCE(NULLIF($4,''),'pen'),COALESCE(NULLIF($5,''),'#ef4444'),
-		        COALESCE(NULLIF($6,0),2), COALESCE(NULLIF($7,0),1), $8)
+		        COALESCE(NULLIF($6::numeric,0),2), COALESCE(NULLIF($7::numeric,0),1), $8)
 		ON CONFLICT (id) DO NOTHING`,
 		in.ID, sheetID, userID, in.Tool, in.Color, in.Width, in.Opacity, string(in.Geometry))
 	if err != nil {
